requirements.txt
# Database
neo4j>=5.14.0

# LLM
groq>=0.4.0
langchain-groq>=0.1.0
langchain-core>=0.1.0

# Environment variables
python-dotenv>=1.0.0

# Frontend
streamlit>=1.28.0

# Utilities
pydantic>=2.5.0
